refactor: separate SignUpForm component, add sign up request result if it fails
