+++ src/pages/signup.tsx
@@ -1,75 +1,15 @@
-import { useRouter } from "next/router";
-import Cookie from "js-cookie";
 import { type GetServerSideProps } from "next";
-import { AUTHORIZATION_TOKEN_KEY } from "~/utils/auth/authorizationTokenKey";
 import { AuthLayout } from "~/component/auth/authLayout";
 import Link from "next/link";
-import { AuthForm } from "~/component/auth/authForm";
+import { SignUpForm } from "~/component/auth/signUpForm";
 import { Layout } from "~/component/layout/layout";
 import { withoutAuth } from "~/utils/auth/withoutAuth";
-import type { User } from "~/utils/dbSchema/models";
 
-const Signup = () => {
-  const router = useRouter();
-  type SignupResponse = {
-    message: string;
-    token: string;
-    user: User;
-  };
-
-  const switchToPageAfterSignup = async (
-    signupResponse: SignupResponse
-  ): Promise<void> => {
-    switch (signupResponse.user.userType) {
-      case "EMPLOYER":
-        await router.push("/home/profile");
-        break;
-      case "CANDIDATE":
-        await router.push("/my/profile");
-        break;
-      case "ADMIN":
-        await router.push("/admin");
-        break;
-    }
-  };
-
-  async function createAccount(
-    login: string,
-    password: string,
-    userType: string | undefined
-  ) {
-    try {
-      const response = await fetch("/api/auth/signup", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ login, password, userType }),
-      });
-
-      if (response.ok) {
-        const data = (await response.json()) as SignupResponse;
-
-        Cookie.set(AUTHORIZATION_TOKEN_KEY, data.token, {
-          expires: 30,
-          path: "/",
-        });
-
-        void switchToPageAfterSignup(data);
-      } else {
-        const errorData = (await response.json()) as SignupResponse;
-        console.error(errorData.message);
-      }
-    } catch (error) {
-      console.error("Network error:", error);
-    }
-  }
-
+export default function Signup() {
   return (
     <Layout>
       <AuthLayout authorizationType="Sign up">
-        <AuthForm
-          handleFormSubmit={createAccount}
-          authorizationType="Sign up"
-        />
+        <SignUpForm />
         <p className="mt-10 text-center text-sm text-gray-500">
           Already have an account?{" "}
           <Link
@@ -82,8 +22,6 @@ const Signup = () => {
       </AuthLayout>
     </Layout>
   );
-};
-
-export default Signup;
+}
 
 export const getServerSideProps: GetServerSideProps = withoutAuth();
